requirements.txt
phidata
google-generativeai
duckduckgo-search
streamlit
python-dotenv